// src/types.d.ts
export interface APIChannelMin {
	id: string;
	name: string;
	englishName?: string;
	type: 'vtuber' | 'subber';
	photo?: string;
}
// ...
export interface APIVideo {
	id: string;
	title: string;
	type: 'stream' | 'clip';
	topic_id?: string;
	published_at?: string;
	available_at: string;
	duration: number;
	status: 'new' | 'upcoming' | 'live' | 'past' | 'missing';
	start_scheduled?: string;
	start_actual?: string;
	end_actual?: string;
	live_viewers?: number;
	description: string;
	songCount: number;
	channel_id: string;
}

/**
 * Model corresponding to https://holodex.stoplight.io/docs/holodex/holodex_v2.yaml/components/schemas/VideoWithChannel
 */
export interface APIVideoWithChannel extends APIVideo {
	channel: APIChannelMin;
}

/**
 * Model corresponding to mentions in https://holodex.stoplight.io/docs/holodex/holodex_v2.yaml/components/schemas/VideoFull
 */
export interface APIMentions extends APIChannelMin {
	org?: string;
}

/**
 * Model corresponding to https://holodex.stoplight.io/docs/holodex/holodex_v2.yaml/components/schemas/VideoFull
 */
export interface APIVideoFull extends APIVideo {
	clips?:	APIVideoWithChannel[];
	sources?: APIVideoWithChannel[];
	refers?: APIVideoWithChannel[];
	simulcasts?: APIVideoWithChannel[];
	mentions?: APIMentions[];
	songs?: number;
}
// ...
export interface ChannelMin {
	id: string;
	name: string;
	englishName?: string;
	type: 'vtuber' | 'subber';
	photo?: string;
}

export interface Channel extends ChannelMin {
	org?: string;
	subOrg?: string;
	banner?: string;
	twitter?: string;
	videoCount?: number;
	subscriberCount?: number;
	viewCount?: number;
	clipCount?: number;
	lang?: string;
	publishedAt: Date;
	inactive: boolean;
	description: string;
}

export interface Video {
	id: string;
	title: string;
	type: 'stream' | 'clip';
	topicId?: string;
	publishedAt?: Date;
	availableAt?: Date;
	duration: number;
	status: string;
	startScheduled?: Date;
	startActual?: Date;
	endActual?: Date;
	liveViewers?: number;
	description: string;
	songCount?: number;
	channelId: string;
}

export interface VideoWithChannel extends Video {
	channel: ChannelMin;
}

export interface VideoFull extends Video {
	clips?: VideoWithChannel[];
	sources?: VideoWithChannel[];
	refers?: VideoWithChannel[];
	simulcasts?: VideoWithChannel[];
	mentions?: ChannelMin[];
	songs?: number;
}

export interface Comment {
	commentKey: string;
	videoId: string;
	message: string;
}

// Additional return types
export interface PaginatedChannelVideosData {
	total: number;
	items: VideoFull[];
}

export interface Mention extends ChannelMin {
	org?: string;
}